clean code clinic manager view
