+++ packages/doctor_pet/lib/common_widget/custom_button/custom_button_action_widget.dart
@@ -1,0 +1,41 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
+
+class CustomButtonActionWidget extends StatefulWidget {
+  const CustomButtonActionWidget({
+    Key? key,
+    this.label,
+    this.bgColor,
+    this.btnColor,
+  }) : super(key: key);
+
+  final String? label;
+  final Color? bgColor;
+  final Color? btnColor;
+
+  @override
+  State<CustomButtonActionWidget> createState() => _CustomButtonActionWidgetState();
+}
+
+class _CustomButtonActionWidgetState extends State<CustomButtonActionWidget> {
+  @override
+  Widget build(BuildContext context) {
+    return ElevatedButton.icon(
+        icon: Icon(Icons.add, color: Colors.white),
+        style: ElevatedButton.styleFrom(
+          backgroundColor: widget.bgColor,
+        ),
+        onPressed: () {
+          showDatePicker(
+              context: context,
+              firstDate: (DateTime(1900, 1, 1)),
+              lastDate: (DateTime(2500, 12, 31)));
+        },
+        label: Text(
+          widget.label.toString(),
+          style: TextStyle(
+            color: Colors.white,
+          ),
+        ));
+  }
+}
