+++ packages/doctor_pet/lib/views/clinic_manager/invoice_report/invoice_report_binding.dart
@@ -1,4 +1,4 @@
-import 'package:doctor_pet/views/clinic_manager/invoicereport/invoicereport_controller.dart';
+import 'package:doctor_pet/views/clinic_manager/invoice_report/invoice_report_controller.dart';
 import 'package:get/get.dart';
 
 class InvoiceReportBinding extends Bindings {
