+++ packages/doctor_pet/lib/views/clinic_manager/invoice_report/invoice_report_controller.dart
@@ -1,0 +1,129 @@
+import 'package:doctor_pet/core/data/invoicereport.dart';
+import 'package:get/get.dart';
+
+class InvoiceReportController extends GetxController {
+  final List<invoicereport> dataInvoiceReport = [
+      invoicereport(
+          patientID: '1',
+          patientName: 'One Zi',
+          cost: 5,
+          day: DateTime(2024, 3, 7, 12, 6, 43),
+          prescription: 'Thuoc ho'),
+      invoicereport(
+          patientID: '2',
+          patientName: 'Duck Ant',
+          cost: 7,
+          day: DateTime(2024, 3, 7, 12, 6, 43),
+          prescription: 'Thuoc cam'),
+      invoicereport(
+          patientID: '1',
+          patientName: 'Two Zi',
+          cost: 7,
+          day: DateTime(2024, 3, 7, 12, 6, 43),
+          prescription: 'Thuoc sot'),
+      invoicereport(
+          patientID: '1',
+          patientName: 'oK',
+          cost: 5,
+          day: DateTime(2024, 3, 7, 12, 6, 43),
+          prescription: 'Thuoc tieu chay'),
+      invoicereport(
+          patientID: '1',
+          patientName: 'oK',
+          cost: 5,
+          day: DateTime(2024, 3, 7, 12, 6, 43),
+          prescription: 'Thuoc tieu chay'),
+      invoicereport(
+          patientID: '1',
+          patientName: 'oK',
+          cost: 5,
+          day: DateTime(2024, 3, 7, 12, 6, 43),
+          prescription: 'Thuoc tieu chay'),
+      invoicereport(
+          patientID: '1',
+          patientName: 'oK',
+          cost: 5,
+          day: DateTime(2024, 3, 7, 12, 6, 43),
+          prescription: 'Thuoc tieu chay'),
+      invoicereport(
+          patientID: '1',
+          patientName: 'oK',
+          cost: 5,
+          day: DateTime(2024, 3, 7, 12, 6, 43),
+          prescription: 'Thuoc tieu chay'),
+      invoicereport(
+          patientID: '1',
+          patientName: 'oK',
+          cost: 5,
+          day: DateTime(2024, 3, 7, 12, 6, 43),
+          prescription: 'Thuoc tieu chay'),
+      invoicereport(
+          patientID: '1',
+          patientName: 'oK',
+          cost: 5,
+          day: DateTime(2024, 3, 7, 12, 6, 43),
+          prescription: 'Thuoc tieu chay'),
+      invoicereport(
+          patientID: '34',
+          patientName: 'Duy',
+          cost: 5,
+          day: DateTime(2024, 3, 7, 12, 6, 43),
+          prescription: 'Thuoc tieu chay'),
+      invoicereport(
+          patientID: '123',
+          patientName: 'Dep trai',
+          cost: 5,
+          day: DateTime(2024, 3, 7, 12, 6, 43),
+          prescription: 'Thuoc tieu chay'),
+      invoicereport(
+          patientID: '13',
+          patientName: 'Nguyen Van A',
+          cost: 5,
+          day: DateTime(2024, 3, 7, 12, 6, 43),
+          prescription: 'Thuoc tieu chay'),
+      invoicereport(
+          patientID: '1',
+          patientName: 'Tran Thi B',
+          cost: 5,
+          day: DateTime(2024, 3, 7, 12, 6, 43),
+          prescription: 'Thuoc tieu chay'),
+      invoicereport(
+          patientID: '1',
+          patientName: 'oK',
+          cost: 5,
+          day: DateTime(2024, 3, 7, 12, 6, 43),
+          prescription: 'Thuoc tieu chay'),
+      invoicereport(
+          patientID: '1',
+          patientName: 'One Zi',
+          cost: 5,
+          day: DateTime(2024, 3, 7, 12, 6, 43),
+          prescription: 'Thuoc tieu chay'),
+      invoicereport(
+          patientID: '1',
+          patientName: 'Nguyen Van Dep Trai',
+          cost: 5,
+          day: DateTime(2024, 3, 7, 12, 6, 43),
+          prescription: 'Thuoc tieu chay'),
+      invoicereport(
+          patientID: '1',
+          patientName: 'oK',
+          cost: 5,
+          day: DateTime(2024, 3, 7, 12, 6, 43),
+          prescription: 'Thuoc tieu chay'),
+      invoicereport(
+          patientID: '1',
+          patientName: 'oK',
+          cost: 5,
+          day: DateTime(2024, 3, 7, 12, 6, 43),
+          prescription: 'Thuoc tieu chay'),
+      invoicereport(
+          patientID: '1',
+          patientName: 'oK',
+          cost: 5,
+          day: DateTime(2024, 3, 7, 12, 6, 43),
+          prescription: 'Thuoc tieu chay'),
+
+      // Add more pets as needed
+    ];
+}
