+++ packages/doctor_pet/lib/views/clinic_manager/medicine_report/medicinereport_binding.dart
@@ -1,4 +1,4 @@
-import 'package:doctor_pet/views/clinic_manager/medicinereport/medicinereport_controller.dart';
+import 'package:doctor_pet/views/clinic_manager/medicine_report/medicinereport_controller.dart';
 import 'package:get/get.dart';
 
 class MedicineReportBinding extends Bindings {
