+++ packages/doctor_pet/lib/views/clinic_manager/medicine_report/medicinereport_controller.dart
@@ -1,0 +1,51 @@
+import 'package:doctor_pet/core/data/medicinereport.dart';
+import 'package:get/get.dart';
+
+class MedicineReportController extends GetxController {
+  final List<medicinereport> dataMedicneReport = [
+    medicinereport(
+        medicineName: 'Thuoc ho', unitPrice: 6, quantity: 5, totalPrice: 6),
+    medicinereport(
+        medicineName: 'Thuoc ho', unitPrice: 6, quantity: 5, totalPrice: 6),
+    medicinereport(
+        medicineName: 'Thuoc ho', unitPrice: 6, quantity: 5, totalPrice: 6),
+    medicinereport(
+        medicineName: 'Thuoc ho', unitPrice: 6, quantity: 5, totalPrice: 6),
+    medicinereport(
+        medicineName: 'Thuoc ho', unitPrice: 6, quantity: 5, totalPrice: 6),
+    medicinereport(
+        medicineName: 'Thuoc ho', unitPrice: 6, quantity: 5, totalPrice: 6),
+    medicinereport(
+        medicineName: 'Thuoc ho', unitPrice: 6, quantity: 5, totalPrice: 6),
+    medicinereport(
+        medicineName: 'Thuoc ho', unitPrice: 6, quantity: 5, totalPrice: 6),
+    medicinereport(
+        medicineName: 'Thuoc ho', unitPrice: 6, quantity: 5, totalPrice: 6),
+    medicinereport(
+        medicineName: 'Thuoc ho', unitPrice: 6, quantity: 5, totalPrice: 6),
+    medicinereport(
+        medicineName: 'Thuoc ho', unitPrice: 6, quantity: 5, totalPrice: 6),
+    medicinereport(
+        medicineName: 'Thuoc ho', unitPrice: 6, quantity: 5, totalPrice: 6),
+    medicinereport(
+        medicineName: 'Thuoc ho', unitPrice: 6, quantity: 5, totalPrice: 6),
+    medicinereport(
+        medicineName: 'Thuoc ho', unitPrice: 6, quantity: 5, totalPrice: 6),
+    medicinereport(
+        medicineName: 'Thuoc ho', unitPrice: 6, quantity: 5, totalPrice: 6),
+    medicinereport(
+        medicineName: 'Thuoc ho', unitPrice: 6, quantity: 5, totalPrice: 6),
+    medicinereport(
+        medicineName: 'Thuoc ho', unitPrice: 6, quantity: 5, totalPrice: 6),
+    medicinereport(
+        medicineName: 'Thuoc ho', unitPrice: 6, quantity: 5, totalPrice: 6),
+    medicinereport(
+        medicineName: 'Thuoc ho', unitPrice: 6, quantity: 5, totalPrice: 6),
+    medicinereport(
+        medicineName: 'Thuoc ho', unitPrice: 6, quantity: 5, totalPrice: 6),
+    medicinereport(
+        medicineName: 'Thuoc ho', unitPrice: 6, quantity: 5, totalPrice: 6),
+
+    // Add more pets as needed
+  ];
+}
