--- packages/doctor_pet/lib/views/clinic_manager/medicinereport/medicinereport_controller.dart
@@ -1,5 +1,0 @@
-import 'package:get/get.dart';
-
-class MedicineReportController extends GetxController {
-  
-}
